requirements.txt
fastapi
uvicorn
langdetect
google-cloud-texttospeech